+++ src/util/index.ts
@@ -2,3 +2,5 @@ export * from "./Functions";
 export { default as Types } from "./Types";
 export { default as Validations } from "./Validations";
 export { default as Assertions } from "./Assertions";
+export { default as Generator } from "./Generator";
+export { default as Properties } from "./Properties";